collision things

+++ CESystem/MouseInput.h
@@ -7,7 +7,7 @@ class MouseInput
 private:
 	double time;
 public:
-	void runSystem(std::vector<Entity*>& entityList, sf::Vector2i mousePosition, double time);
+	void runSystem(std::vector<Entity*>& entityList, sf::Vector2i mousePosition, float time);
 
 	MouseInput();
 	~MouseInput();
